refactor(Car): Extract Car package in domain
- 도메인 내 자동차 관련 패키지 분리
