+++ src/main/java/racingcar/domain/car/Engine.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 @FunctionalInterface
 public interface Engine {
